Moved computerarchitectureinfo+interface out to it's own file.
