+++ src/ImageProcessor/Imaging/ComputerArchitectureInfo.cs
@@ -1,0 +1,12 @@
+namespace ImageProcessor.Imaging
+{
+    using System;
+
+    public class ComputerArchitectureInfo : IComputerArchitectureInfo
+    {
+        public bool IsLittleEndian()
+        {
+            return BitConverter.IsLittleEndian;
+        }
+    }
+}
