+++ src/ImageProcessor/Imaging/IComputerArchitectureInfo.cs
@@ -1,0 +1,7 @@
+namespace ImageProcessor.Imaging
+{
+    public interface IComputerArchitectureInfo
+    {
+        bool IsLittleEndian();
+    }
+}
